backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
sse-starlette>=1.8.0
redis>=5.0.0

# Data Processing & ML
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
prophet>=1.1.4
rapidfuzz>=3.0.0

# AI/LLM (for AI Assistant)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0
langchain-nvidia-ai-endpoints>=0.3.0
python-dotenv>=1.0.0
